requirements.txt
Django==5.1.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.0
psycopg2==2.9.9
